requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anthropic==0.3.6
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
dataclasses-json==0.5.13
distro==1.8.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
langchain==0.0.247
langsmith==0.0.15
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.12
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
tenacity==8.2.2
tokenizers==0.13.3
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
